requirements.txt
flask
requests
flask
requests
psycopg2-binary
